backend_service/news_pipeline/news_scraper.py
from newspaper import Article


def extract_news(url):
    try:
        article = Article(url)
        article.download()
        article.parse()
        return article.text
    except Exception as e:
        print(e)
